optimiser partie investment portfolio
