+++ lib/features/auth/presentation/pages/dashboard/widgets/investment_portfolio/empty_portfolio_widget.dart
@@ -1,0 +1,116 @@
+import 'package:flutter/material.dart';
+import 'package:the_boost/core/constants/colors.dart';
+import 'package:the_boost/core/constants/dimensions.dart';
+
+class EmptyPortfolioWidget extends StatelessWidget {
+  final String currentDate;
+  final String userName;
+
+  const EmptyPortfolioWidget({
+    Key? key,
+    required this.currentDate,
+    required this.userName,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(AppDimensions.paddingL),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(AppDimensions.radiusL),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          // Header section with title and date
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              const Text(
+                "Your Land Tokens",
+                style: TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.bold,
+                  color: AppColors.textPrimary,
+                ),
+              ),
+              Container(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 12,
+                  vertical: 6,
+                ),
+                decoration: BoxDecoration(
+                  color: AppColors.backgroundGreen,
+                  borderRadius: BorderRadius.circular(16),
+                ),
+                child: Row(
+                  children: [
+                    const Icon(
+                      Icons.update,
+                      color: AppColors.primary,
+                      size: 16,
+                    ),
+                    const SizedBox(width: 4),
+                    Text(
+                      currentDate,
+                      style: const TextStyle(
+                        fontSize: 12,
+                        color: AppColors.primary,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+          const SizedBox(height: AppDimensions.paddingL),
+          Icon(
+            Icons.account_balance,
+            color: Colors.grey[400],
+            size: 40,
+          ),
+          const SizedBox(height: AppDimensions.paddingM),
+          const Text(
+            "No investments yet",
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              fontSize: 16,
+            ),
+          ),
+          const SizedBox(height: AppDimensions.paddingS),
+          Text(
+            "Start investing in properties to build your portfolio",
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              color: Colors.grey[600],
+              fontSize: 14,
+            ),
+          ),
+          const SizedBox(height: AppDimensions.paddingL),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.pushNamed(context, '/invest');
+              
+              print('[$currentDate] InvestmentPortfolio: 🔄 Navigating to invest page'
+                  '\n└─ User: $userName');
+            },
+            style: ElevatedButton.styleFrom(
+              backgroundColor: AppColors.primary,
+              foregroundColor: Colors.white,
+            ),
+            child: const Text("Discover Properties"),
+          ),
+        ],
+      ),
+    );
+  }
+}
