+++ lib/features/auth/presentation/pages/dashboard/widgets/investment_portfolio/error_portfolio_widget.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+import 'package:the_boost/core/constants/colors.dart';
+import 'package:the_boost/core/constants/dimensions.dart';
+
+class ErrorPortfolioWidget extends StatelessWidget {
+  final String message;
+  final VoidCallback onRetry;
+  final String currentDate;
+  final String userName;
+
+  const ErrorPortfolioWidget({
+    Key? key,
+    required this.message,
+    required this.onRetry,
+    required this.currentDate,
+    required this.userName,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(AppDimensions.paddingL),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(AppDimensions.radiusL),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withOpacity(0.05),
+            blurRadius: 10,
+            offset: const Offset(0, 2),
+          ),
+        ],
+      ),
+      child: Column(
+        children: [
+          const Icon(
+            Icons.error_outline,
+            color: Colors.orange,
+            size: 40,
+          ),
+          const SizedBox(height: AppDimensions.paddingS),
+          const Text(
+            "Couldn't load your investments",
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              fontSize: 16,
+            ),
+          ),
+          const SizedBox(height: AppDimensions.paddingXS),
+          Text(
+            message,
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              color: Colors.grey[600],
+              fontSize: 14,
+            ),
+          ),
+          const SizedBox(height: AppDimensions.paddingM),
+          TextButton.icon(
+            onPressed: () {
+              onRetry();
+              
+              print('[$currentDate] InvestmentPortfolio: 🔄 Retrying to load tokens'
+                  '\n└─ User: $userName');
+            },
+            icon: const Icon(Icons.refresh),
+            label: const Text("Try Again"),
+            style: TextButton.styleFrom(
+              foregroundColor: AppColors.primary,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
