+++ lib/features/auth/presentation/pages/dashboard/widgets/investment_portfolio/land_info_section_widget.dart
@@ -1,0 +1,186 @@
+import 'package:flutter/material.dart';
+import 'package:the_boost/core/constants/colors.dart';
+import 'package:the_boost/core/constants/dimensions.dart';
+import 'package:the_boost/features/auth/domain/entities/token.dart';
+
+class LandInfoSectionWidget extends StatelessWidget {
+  final int landId;
+  final String landTitle;
+  final String landLocation;
+  final String? landImageUrl;
+  final int tokenCount;
+  final bool anyListed;
+  final List<Token> listedTokens;
+  final String formattedTotalValue;
+  final double profitLoss;
+  final String formattedPercentage;
+
+  const LandInfoSectionWidget({
+    Key? key,
+    required this.landId,
+    required this.landTitle,
+    required this.landLocation,
+    required this.landImageUrl,
+    required this.tokenCount,
+    required this.anyListed,
+    required this.listedTokens,
+    required this.formattedTotalValue,
+    required this.profitLoss,
+    required this.formattedPercentage,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: () {
+        Navigator.pushNamed(
+          context,
+          '/land',
+          arguments: landId,
+        );
+      },
+      child: Padding(
+        padding: const EdgeInsets.all(AppDimensions.paddingM),
+        child: Row(
+          children: [
+            _buildLandImage(landImageUrl),
+            const SizedBox(width: AppDimensions.paddingM),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    landTitle,
+                    style: const TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 16,
+                    ),
+                    maxLines: 1,
+                    overflow: TextOverflow.ellipsis,
+                  ),
+                  Row(
+                    children: [
+                      const Icon(
+                        Icons.location_on,
+                        size: 12,
+                        color: Colors.grey,
+                      ),
+                      const SizedBox(width: 2),
+                      Expanded(
+                        child: Text(
+                          landLocation,
+                          style: TextStyle(
+                            color: Colors.grey[600],
+                            fontSize: 14,
+                          ),
+                          maxLines: 1,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                      ),
+                    ],
+                  ),
+                  const SizedBox(height: 4),
+                  Row(
+                    children: [
+                      Text(
+                        "$tokenCount tokens owned",
+                        style: TextStyle(
+                          color: Colors.grey[600],
+                          fontSize: 12,
+                        ),
+                      ),
+                      if (anyListed)
+                        Container(
+                          margin: const EdgeInsets.only(left: 8),
+                          padding: const EdgeInsets.symmetric(
+                            horizontal: 6,
+                            vertical: 2,
+                          ),
+                          decoration: BoxDecoration(
+                            color: Colors.orange.withOpacity(0.1),
+                            borderRadius: BorderRadius.circular(4),
+                          ),
+                          child: Text(
+                            "${listedTokens.length} Listed",
+                            style: const TextStyle(
+                              color: Colors.orange,
+                              fontSize: 10,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ),
+                    ],
+                  ),
+                ],
+              ),
+            ),
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.end,
+              children: [
+                Text(
+                  formattedTotalValue,
+                  style: const TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 16,
+                  ),
+                ),
+                Row(
+                  children: [
+                    Icon(
+                      profitLoss >= 0
+                          ? Icons.arrow_upward
+                          : Icons.arrow_downward,
+                      color: profitLoss >= 0 ? Colors.green : Colors.red,
+                      size: 16,
+                    ),
+                    Text(
+                      formattedPercentage,
+                      style: TextStyle(
+                        color: profitLoss >= 0 ? Colors.green : Colors.red,
+                        fontSize: 14,
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget _buildLandImage(String? imageUrl) {
+    if (imageUrl != null && imageUrl.isNotEmpty) {
+      return ClipRRect(
+        borderRadius: BorderRadius.circular(8),
+        child: Image.network(
+          imageUrl,
+          width: 50,
+          height: 50,
+          fit: BoxFit.cover,
+          errorBuilder: (context, error, stackTrace) {
+            return _buildPlaceholderImage();
+          },
+        ),
+      );
+    }
+    return _buildPlaceholderImage();
+  }
+
+  Widget _buildPlaceholderImage() {
+    return Container(
+      width: 50,
+      height: 50,
+      decoration: BoxDecoration(
+        color: AppColors.primary.withOpacity(0.1),
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: const Icon(
+        Icons.home,
+        color: AppColors.primary,
+        size: 24,
+      ),
+    );
+  }
+}
